requirements.txt
discord.py==2.1.0
python-dotenv==0.21.0
beautifulsoup4==4.11.1
requests==2.28.1
